requirements.txt
google-generativeai
python-dotenv
langchain
langchain-google-genai
pydantic
requests>=2.31.0
